revert rename netpyne-ui
